+++ R/utils.R
@@ -7,3 +7,8 @@ stringify_date <- function(d) {
     d
   }
 }
+#' Combine 2 vectors by alternating between elements
+#' @noRd
+interleave <- function(v1, v2) {
+  c(rbind(v1, v2))
+}